Moves command classes into moderation.auto and moderation.manual packages for better organization
