+++ .gitignore
@@ -98,7 +98,7 @@ out/
 # Windows image file caches
 Thumbs.db
 
-# Folder config file
+# Folder prepareConfig file
 Desktop.ini
 
 ######################
