+++ src/main/java/moderator/moderation/auto/commands/AntiSpamFilter.java
@@ -1,4 +1,4 @@
-package moderator.moderation;
+package moderator.moderation.auto.commands;
 
 import net.dv8tion.jda.api.entities.*;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
